pyproject.toml
[project]
name = "research_agent_crew"
version = "0.1.0"
description = "research_agent_crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.8.0",
    "fastapi>=0.128.0",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "litellm>=1.75.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.2.2",
]

[project.scripts]
research_agent_crew = "research_agent_crew.main:run"
run_crew = "research_agent_crew.main:run"
train = "research_agent_crew.main:train"
replay = "research_agent_crew.main:replay"
test = "research_agent_crew.main:test"
run_with_trigger = "research_agent_crew.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
